// frontend/src/types.ts
export interface Notice {
  type: "success" | "error" | "warning";
  message: string;
}